Move upload functionality to plain object
